New regression test for long doubles.
git-svn-id: https://repo.eecs.berkeley.edu/svn/users/ksen/projects/corvette-instrumentor@5803 1c3fd406-9068-4342-8353-511fcdf43622

+++ tests/array_heap_long.c
@@ -3,14 +3,14 @@
 #include <stdio.h> 
 
 int main() {
-  double* array = (double*) malloc(5*sizeof(double));
+  long double* array = (long double*) malloc(5*sizeof(long double));
 
   int i = 0;
   for(; i < 5; i++) {
     array[i] = i + 0.3;
   }
 
-  double result = array[3];
+  long double result = array[3];
 
   return 0;
 }
